pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "ssid-open-core"
version = "0.1.0"
description = "SSID Open-Core: public libraries, SDKs, and shared types"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
dependencies = [
    "jsonschema>=4.23.0",
    "pyyaml>=6.0.2",
]

[project.optional-dependencies]
test = [
    "pytest>=8.3.0",
]
dev = [
    "ruff>=0.9.0",
]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]
ignore = [
    "E501",   # line too long (handled by formatter)
    "E402",   # module level import not at top (common in scripts with sys.path)
    "E741",   # ambiguous variable name (l, O, I) -- common in comprehensions
    "B008",   # function call in default argument (FastAPI Depends pattern)
    "SIM102", # collapsible if -- readability preference
    "SIM105", # contextlib.suppress -- readability preference
]

[tool.ruff.lint.isort]
known-first-party = ["ssid_open_core"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
